Moved to unbounded chunk-lists
The previous allocator used a finite set of chunk-lists. This had the
nice garuntee of strict constant allocation cost, but incurred
unnecessary fragmentation once the set was exceeded.

By using a linked-list of chunk-lists, this fragmentation can be avoided
and the chunk-list scan potenially short-circuited, at the downside of
introducing the quantity of different allocation sizes as a variable in
the allocation cost.