Added EVENTS_EVENT_SIZE define for calculated queue allocations
In the events library, the structure of an event is correctly not
part of the public api. Unfortunately, this makes it difficult to
calculate the space needed for a finite set of events.

EVENTS_EVENT_SIZE is defined as the space needed for the smallest
event, the event allocated in event_call.

This allows queues to be created based on a number of events:
equeue_create(&q, 12 * (EVENTS_EVENT_SIZE+sizeof(struct context)));