Added the ability to background an event queue onto a single-shot timer
The equeue_background function requires only a user-provided update
function, which is called to indicate when the queue should be
dispatched.

The equeue_background function adds a new level of composability to the
event queue library. Allowing equeues to be easily backgrounded on
hardware timers or even be combined with existing event loops.